moving Menu component and stylesheet
